Part 1 complete

+++ src/main/java/uk/ac/ed/inf/powergrab/Position.java
@@ -28,14 +28,14 @@ public class Position {
 		//    --if a point is within the square or not
 		//bl = bottom left
 		//tr = top right
-		Float bl_lat  = 55.942617;
-		Float bl_long = -3.192473;
-		Float tr_lat  = 55.946233;
-		Float tr_long = -3.184319;
+		double bl_lat  = 55.942617;
+		double bl_long = -3.192473;
+		double tr_lat  = 55.946233;
+		double tr_long = -3.184319;
 		
 		//new variables for the current position to make the algorithm more readable
-        Float lat = this.latitude;
-        Float lon = this.longitude;
+        double lat = this.latitude;
+        double lon = this.longitude;
 
         if (lat > bl_lat && lat < tr_lat && lon > bl_long && lon < tr_long ){
             return true;
